feat: add like functionality to comments and update UI for better interaction

+++ dollog/backend/src/main/java/com/doldev/dollog/domain/comment/entity/Comment.java
@@ -53,7 +53,7 @@ public class Comment {
     private List<Comment> child = new ArrayList<>();
 
     @ColumnDefault("0")
-    private Long likeCnt;
+    private int likeCnt;
 
     @Builder.Default
     @OneToMany(mappedBy = "comment", cascade = CascadeType.REMOVE)
